fix(process): move Enabled() to top-level agent package to break import cycle
The agent/impl tests import runnerfx which imports runner/impl which
previously imported agentimpl for Enabled(), creating a cycle. Moving
Enabled(), Once, and enabledHelper() to the top-level comp/process/agent
package (package agent) lets both agent/impl and runner/impl import it
without creating any cycle.
